fix specs

+++ app/services/events/stores/aggregated_clickhouse/weighted_sum_query.rb
@@ -48,7 +48,7 @@ module Events
               timestamp,
               difference,
               SUM(difference) OVER (ORDER BY timestamp ASC ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW) AS cumul,
-              date_diff('seconds', timestamp, leadInFrame(timestamp, 1, toDateTime64(:to_datetime, 5, 'UTC')) OVER (ORDER BY timestamp ASC ROWS BETWEEN CURRENT ROW AND UNBOUNDED FOLLOWING)) AS second_duration,
+              date_diff('seconds', timestamp, leadInFrame(timestamp, 1, toDateTime64(:to_datetime, 5, 'UTC')) OVER (ORDER BY timestamp ASC ROWS BETWEEN CURRENT ROW AND UNBOUNDED FOLLOWING))::INT AS second_duration,
               (#{period_ratio_sql}) AS period_ratio
             FROM events_data
             ORDER BY timestamp ASC
